Moved actions into separate source file.
